[REFACTOR] - All main styles files created - Time to CODE!
