+++ package.json
@@ -8,7 +8,6 @@
     "dev": "webpack-dev-server --mode development",
     "server": "webpack --mode development --open && npx serve dist -s",
     "build": "webpack --mode production"
-
   },
   "keywords": [
     "javascript",
@@ -26,6 +25,7 @@
     "html-loader": "^1.3.0",
     "html-webpack-plugin": "^4.4.1",
     "node-sass": "^4.14.1",
+    "resolve-url-loader": "^3.1.1",
     "sass-loader": "^10.0.2",
     "webpack": "^4.44.1",
     "webpack-cli": "^3.3.12",
