+++ src/assets/styles/App.scss
@@ -1,37 +1,18 @@
-@import url("https://fonts.googleapis.com/css2?family=Roboto:wght@300;400;500;700&display=swap");
-@import url("https://fonts.googleapis.com/css2?family=Lato:wght@300;400;700&family=Roboto:wght@300;400;500;700&display=swap");
+@import './partials/fonts.scss';
+@import './reboot.scss';
 
-html {
-  box-sizing: border-box;
-  font-family: "Roboto", "Lato";
-}
-*,
-*:before,
-*:after {
-  box-sizing: inherit;
-}
-ul,
-li,
-h1,
-h2,
-h3,
-p,
-button {
-  margin: 0;
-  padding: 0;
-}
-ul {
-  list-style: none;
-}
-button {
-  background: transparent;
-  border: 0;
-  outline: 0;
-}
-body {
-  background: #fefefe;
-  height: 100vh;
-  margin: 0 auto;
-  overscroll-behavior: none;
-  width: 100%;
+@import './Media.scss';
+@import './Vars.scss';
+
+@import './partials/spacing.scss';
+@import './partials/colors.scss';
+@import './partials/shadows.scss';
+@import './partials/text.scss';
+
+
+:root {
+  --color-primary: $color-primary;
+  --color-black: $color-black;
+  --color-white: $color-white;
+  --color-background: $color-background;
 }
