+++ src/assets/styles/partials/colors.scss
@@ -1,0 +1,12 @@
+$color-primary: hsl(158,69%,49%);
+$color-primary-90: hsla(158,69%,49%,.9);
+$color-primary-60: hsla(158,69%,49%,.6);
+$color-primary-30: hsla(158,69%,49%,.3);
+
+$color-black: hsl(240,27%,12%);
+$color-black-90: hsla(240,27%,12%,.9);
+$color-black-60: hsla(240,27%,12%,.6);
+$color-black-30: hsla(240,27%,12%,.3);
+
+$color-white: hsl(152,59%,94%);
+$color-background: hsl(0,0%,99%);
