+++ src/assets/styles/partials/shadows.scss
@@ -1,0 +1,7 @@
+$shadow-small: 0 0 12px 3px hsla(240,27%,12%, .12);
+$shadow-normal: 0 0 18px 3px hsla(240,27%,12%, .15);
+$shadow-large: 0 0 21px 3px hsla(240,27%,12%, .18);
+
+$shadow-small-green: 0 0 12px 3px hsla(158,69%,49%, .12);
+$shadow-normal-green: 0 0 18px 3px hsla(158,69%,49%, .15);
+$shadow-large-green: 0 0 21px 3px hsla(158,69%,49%, .18);
