+++ src/assets/styles/partials/sizes.scss
@@ -1,0 +1,6 @@
+$font-size-large: 66px;
+$font-size-big: 60px;
+$font-size-medium: 42px;
+$font-size-normal: 30px;
+$font-size-small: 21px;
+$font-size-extra-small: 15px;
