+++ src/assets/styles/partials/spacing.scss
@@ -1,0 +1,7 @@
+$margin-extra-small: 3px;
+$margin-small: 6px;
+$margin-normal: 12px;
+$margin-big: 18px;
+$margin-large: 21px;
+$margin-extra-big: 30px;
+$margin-extra-large: 60px;
