+++ src/assets/styles/reboot.scss
@@ -1,0 +1,34 @@
+html {
+  box-sizing: border-box;
+  font-family: $font-secondary, "Roboto", "Lato";
+}
+*,
+*:before,
+*:after {
+  box-sizing: inherit;
+}
+ul,
+li,
+h1,
+h2,
+h3,
+p,
+button {
+  margin: 0;
+  padding: 0;
+}
+ul {
+  list-style: none;
+}
+button {
+  background: transparent;
+  border: 0;
+  outline: 0;
+}
+body {
+  background: #fefefe;
+  height: 100vh;
+  margin: 0 auto;
+  overscroll-behavior: none;
+  width: 100%;
+}
