+++ src/index.jsx
@@ -1,9 +1,11 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./components/App";
 import { Provider } from "react-redux";
-import reducer from "./reducers/index";
 import { createStore } from "redux";
+import reducer from "./reducers/index";
+
+import './assets/styles/app.scss';
+import App from "./routes/App";
 
 // Global state
 const initialState = {
